Setup more variables for more modularity
Lot's of folder and files are define as variables, and could be
overwritten in the user/host files. This make the configuration more
modular or at least more easy to use on different platform and hosts.

For now this is used almost only for org-mode but this is a start.